Remove leader election capability
Since the watch is set up using the hub cluster config, the leader
election code attempts to retrieve the config map in the local
namespace. This results in insufficient privileges since the hub cluster
config is configured with the RBAC permissions to operate in the hub
cluster and not the spoke cluster.